chore: parse animator even if trace and optimize is not activated
